Move ReportPeriod enum to a new file
