+++ src/components/engagement/engagement.resolver.ts
@@ -13,7 +13,7 @@ import {
   PeriodicReportService,
   SecuredPeriodicReportList,
 } from '../periodic-report';
-import { ReportType } from '../periodic-report/dto/type.enum';
+import { ReportType } from '../periodic-report/dto';
 import {
   CreateInternshipEngagementInput,
   CreateInternshipEngagementOutput,
