+++ src/components/periodic-report/dto/index.ts
@@ -4,3 +4,5 @@ export * from './create-periodic-report.dto';
 export * from './update-periodic-report.dto';
 export * from './list-periodic-reports.dto';
 export * from './upload-periodic-report.dto';
+export * from './report-period.enum';
+export * from './report-type.enum';
