+++ src/components/periodic-report/dto/report-period.enum.ts
@@ -1,18 +1,11 @@
 import { ObjectType, registerEnumType } from '@nestjs/graphql';
 import { SecuredEnum, SecuredProperty } from '../../../common';
 
-export enum ReportType {
-  Financial = 'Financial',
-  Progress = 'Progress',
-  Narrative = 'Narrative',
-}
-
 export enum ReportPeriod {
   Monthly = 'Monthly',
   Quarterly = 'Quarterly',
 }
 
-registerEnumType(ReportType, { name: 'ReportType' });
 registerEnumType(ReportPeriod, { name: 'PeriodType' });
 
 @ObjectType({
