+++ src/components/periodic-report/dto/report-type.enum.ts
@@ -1,0 +1,9 @@
+import { registerEnumType } from '@nestjs/graphql';
+
+export enum ReportType {
+  Financial = 'Financial',
+  Progress = 'Progress',
+  Narrative = 'Narrative',
+}
+
+registerEnumType(ReportType, { name: 'ReportType' });
