+++ src/components/project/dto/create-project.dto.ts
@@ -13,7 +13,7 @@ import {
   NameField,
   Sensitivity,
 } from '../../../common';
-import { ReportPeriod } from '../../periodic-report/dto/type.enum';
+import { ReportPeriod } from '../../periodic-report/dto';
 import { IProject, Project } from './project.dto';
 import { ProjectStep } from './step.enum';
 import { ProjectType } from './type.enum';
