+++ src/components/project/dto/project.dto.ts
@@ -24,7 +24,7 @@ import { Directory } from '../../file/dto';
 import { SecuredTags } from '../../language/dto/language.dto';
 import { Location } from '../../location/dto';
 import { Partnership } from '../../partnership/dto';
-import { SecuredReportPeriod } from '../../periodic-report/dto/type.enum';
+import { SecuredReportPeriod } from '../../periodic-report/dto';
 import { Pinnable } from '../../pin/dto';
 import { ProjectMember } from '../project-member/dto';
 import { ProjectStatus } from './status.enum';
